requirements.txt
streamlit==1.28.1
google-generativeai==0.3.2
Pillow==10.0.1
python-dotenv==1.0.0
pydantic==2.5.0
typing-extensions==4.8.0
requests==2.31.0
opencv-python==4.8.1.78
numpy==1.24.3
